v1/requirements.txt
pandas
numpy
pyarrow
scikit-learn
category-encoders
scipy
Unidecode
PyYAML
flaml
lightgbm
xgboost
catboost
joblib
